Move modify commands to `Modifys` module
Was `Modifiers` but this is inconsistent with `Deletes` and `Shows` etc.
