+++ lib/commands/modifys/groups.rb
@@ -20,36 +20,41 @@
 # https://github.com/alces-software/inventoryware
 #==============================================================================
 
+#TODO extract shared `Modifys` code to top level 'modify.rb' command file
 module Inventoryware
   module Commands
-    module Modifiers
-      class Location < Command
+    module Modifys
+      class Groups < Command
         def run
-          other_args = []
+          other_args = ["group"]
           nodes = Utils::resolve_node_options(@argv, @options, other_args)
-          node_locations = Utils::select_nodes(nodes, @options)
 
-          fields = {
-            'site' => {'name' => nil, 'value' => nil},
-            'room' => {'name' => nil, 'value' => nil},
-            'rack' => {'name' => nil, 'value' => nil},
-            'start_unit' => {'name' => 'starting rack unit', 'value' => nil},
-          }
-
-          # Get input REPL style
-          fields.each do |field, hash|
-            name = hash['name'] ? hash['name'] : field
-            value = ask("Enter a #{name} or press enter to skip")
-            hash['value'] = value unless value == ''
+          if @options.primary and @options.remove
+            raise ArgumentError, <<-ERROR
+  Cannot remove a primary group
+            ERROR
           end
 
-          # save data
+          #TODO DRY up? group is defined twice
+          group = @argv[0]
+
+          node_locations = Utils::select_nodes(nodes,
+                                               @options,
+                                               return_missing = true)
+
           node_locations.each do |location|
             node_data = Utils::read_node_or_create(location)
-            fields.each do |field, hash|
-              if hash['value']
-                node_data['mutable'][field] = hash['value']
+            if @options.primary
+              node_data['mutable']['primary_group'] = group
+            else
+              sec = node_data['mutable'].fetch('secondary_groups', nil)&.split(',')
+              if @options.remove and sec.include?(group)
+                sec.delete(group)
+              elsif not @options.remove
+                sec ? sec << group : sec = [group]
+                sec.uniq!
               end
+              node_data['mutable']['secondary_groups'] = sec.join(',')
             end
             Utils::output_node_yaml(node_data, location)
           end
